requirements.txt
goodreads==0.3.2
python-dotenv==0.13.0
